auth_services/requirements.txt
fastapi==0.95.0
boto3==1.26.0
python-dotenv==0.21.0
uvicorn==0.22.0
